lisää tehtävän 4.10 ja test_helper
