+++ package.json
@@ -23,6 +23,7 @@
   },
   "devDependencies": {
     "cross-env": "^5.2.0",
+    "eslint": "^6.2.2",
     "jest": "^24.9.0",
     "nodemon": "^1.19.1",
     "supertest": "^4.0.2"
