+++ tests/test_helper.js
@@ -1,0 +1,16 @@
+const initBlogs = [
+  {
+    title: 'Blog One',
+    author: 'First Author',
+    url: 'localhost:something',
+    likes: 3
+  },
+  {
+    title: 'Blog Two',
+    author: 'Author Also',
+    url: 'localhost:something/2',
+    likes: 2
+  }
+]
+
+module.exports = initBlogs
